app/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0
Pillow>=9.0.0
numpy<2.0
